moved initialize
